fix: refactor again to not use a class
