+++ CHANGELOG.md
@@ -1,11 +1,10 @@
 # CHANGELOG
 
-## v2.0.0 (TODO)
+## v2.0.0 (2021-11-22)
 
-* Refactored the entire app to create instances of a table and perform actions on them such as `generate` (see README for info)
-* Added missing white and black colors, changed how to invoke the colors (see README for info)
+* Refactored the entire app to be more straightforward to use. Create a new table by calling `pretty_tables.create()` and pass in your arguments
+* Added missing white and black colors, changed how to invoke the colors (now `pretty_tables.Colors.blue` instead of `pretty_tables.TableColors.OKBLUE`)
 * Validate `color` parameter which was previously skipped
-* Various quality of life improvements both for the user and for maintainers of this package
 * Overhauls documentation providing all customization options (closes #2)
 
 ## v1.3.0 (2021-09-10)
